Buycom_backend/api/serializers.py
from rest_framework import serializers
from .models import Login, CompanyDetails, Return, Score, CompanyGSTRecord
from django.contrib.auth.hashers import check_password
from django.contrib.auth import authenticate


class LoginSerializer(serializers.Serializer):
    username = serializers.CharField(max_length=50)
    password = serializers.CharField(write_only=True)

    def validate(self, attrs):
        username = attrs.get('username')
        password = attrs.get('password')

        user = authenticate(username=username, password=password)

        if user is None:
            raise serializers.ValidationError("Invalid credentials")
        
        return attrs

class CompanyDetailsSerializer(serializers.ModelSerializer):
    class Meta:
        model = CompanyDetails
        fields = '__all__'

class ReturnSerializer(serializers.ModelSerializer):
    class Meta:
        model = Return
        fields = '__all__'

class ScoreSerializer(serializers.ModelSerializer):
    class Meta:
        model = Score
        fields = '__all__'
        
class CompanyGSTRecordSerializer(serializers.ModelSerializer):
    class Meta:
        model = CompanyGSTRecord
        fields = '__all__'
        