Buycom_backend/requirements.txt
antiorm==1.2.1
asgiref==3.8.1
asttokens==2.4.1
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
db==0.1.1
db-sqlite3==0.0.1
debugpy==1.8.7
decorator==5.1.1
dj-database-url==2.3.0
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
executing==2.1.0
fonttools==4.54.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.9
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
pywin32==308
pyzmq==26.2.0
requests==2.32.3
seaborn==0.13.2
serverless-wsgi==3.0.4
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
stripe==11.0.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
wcwidth==0.2.13
Werkzeug==3.0.4
whitenoise==6.7.0
